requirements.txt
Flask==2.2.5
openai
flask_cors
python-dotenv
requests
gunicorn
